Support fail-fast in project pipelines
In some cases like resource constrained environments it is beneficial
to report on changes in a fail fast manner to immediately report if
one job fails. This can be useful especially if a project has many
expensive long-running jobs. This introduces a fail-fast flag in the
project pipeline that let's the project choose the trade off between
full information and quick feedback.

Change-Id: Ie4a5ac8e025362dbaacd3ae82f2e8369f7447a62