Update patch set 2
Patch Set 2:

Do we really want to remove support for old ansibles as soon as ansible EOLs them? They remain on pypi and zuul provides an upgrade path should users need to do that for some other reason. If we do this then zuul is going to be just as annoying to deal with as ansible itself when it comes to ansible breaking things every release.

Patch-set: 2
Label: Verified=0