Update patch set 2
Patch Set 2:

I think we want to remove EOL releases of ansible from zuul, but maybe not this fast.  The question I have is around testing mostly, if we keep this in tree how do we handling testing?

What timeline would be okay for actually removing the code? 6 months, 1 year, when 2.9 comes out?

Patch-set: 2
Label: Workflow=0