Log an error when command socket is missing
If the entrypoint is called as an RPC client, provide a useful error
message when the named pipe it's looking for doesn't exist. This can
be especially useful in situations where a service manager calls the
stop subcommand immediately following a failed start, since merely
raising a connect error can be mistaken for the actual problem.

Change-Id: I821d5e6be29e2068a0fe12b550c094a4b887f40a