Update patch set 1
Patch Set 1:

To solve the original problem, I think we could expose the virtualenv path as an zuul variable, and the ARA roles can use that.  That makes ara special, but that's okay, it is special because it's the thing that needs to run inside Zuul's virtualenv.  That means it's okay for the ara roles to need to know details like that.  Then it makes the other roles able to behave "normally".

Patch-set: 1