Fix memory leak in job cancelation
We were not correctly identifying node requests when canceling jobs
because we were looking for them by job object when they are actually
indexed by job name.  To correct this, use the job name instead.

To reduce the likelihood of similar errors in the future, use the
accessor method, and add type annotations to it and similar methods.

Also, use "job_name" universally in the cancel method as a minor
optimization.

Change-Id: I4051bef64a91eba757674f5534112e9ce6dfdbb6