Update patch set 8
Patch Set 8: Code-Review-1

(1 comment)

a soft -1, but there will be some financial costs to enabling this for zuul.ansible.com, I think we need a global way to disable this or whitelist projects. I'd like users to still have in-tree zuul.yaml, but not change how nodepool nodes are allocated.

Patch-set: 8
Label: Code-Review=-1