Update patch set 8
Patch Set 8:

(1 comment)

Do you mind if we add testing to validate on that? if org/project1 set fail-fast: false in common-project, I would expect org/project1 in tree .zuul.yaml to no be able to change it.

However, my brain seems to think it would be able to override it.

Patch-set: 8