web: honor job dependencies in trigger modal
This change refactors the ProjectVariant page to display a list
of checkbox and to honor job dependencies. When a job requires
another one, both will be added to the buildset.

Moreover, the BuildModal now display all the jobs parameters and
it indicates when a parameter only affects a subset of the jobs.

Change-Id: Ia4b44e2778f7dae68b495f7a0f3d18a373750ec1