WIP: Annotate logs with event id
Currently it's not easy to trace all actions that are triggered by a
specific event. For advanced debugging however this would be very
useful. This change is an attempt to annotate most of the logs with
the event id that triggered this action. This will make it possible to
trace an event through the whole system.

Change-Id: If43cfaebad08c2f2925336337126bd144fe13a7d