Update patch set 5
Patch Set 5: Code-Review+2

Sorry I misremembered -- you are correct this does look like we run docker-compose as root.

Let's rebase the DNM change on this to confirm this fixes the problem.  We could also see if we still need to run docker-compose as root; perhaps we could run it as the zuul user.  If so, that might be a better test.

Patch-set: 5
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2