web: remove SafeLoader left-over from MappingProxyEncoder
When a job has a single key vars (or extra-vars/host-vars) that is named
after a zuul_node_types such as 'job', then the MappingProxy get
a SourceContext and ZuulMark attribute which can not be serialized.
The ZuulSafeLoader shouldn't add such node attributes in the first place,
but it doesn't seems possible to prevent that.

Change-Id: I9fe28a204d41944f86644117c16129234c9d3583