Differentiate between queued and waiting jobs in zuul web UI
When working with job dependencies, the zuul web UI does not
differentiate between jobs that are able to run and those that are
waiting for another job to finish before they can actually run. As both
are marked as 'queued', this often leads to confusion and people are
wondering why the queuing takes so long. To better differentiate between
these states, this change introduces a new state 'waiting' in zuul web.

Change-Id: Id0e85dc2db9660925209dc3cee365aa5c1241190