Log github requests with annotated events
In order to optimize github requests it is important that we can
analyse the request flow triggered by a single event. Therefore add a
response hook to the github session that logs the request along with
some data like event id, request method, response code, response size
and request duration.

Change-Id: I0164e40926ef339ea53af048940934104ba15beb