Read multiple *.conf files from directory
This commit allows for reading multiple .conf files from a directory.
The 2 default directories are `/etc/zuul/conf.d` and `~/.zuul.d`.

In environment like kuernetes, passwords and other sensitive data are
stored in secrets. This commit allows for storing only the section of
the file containing sensitive data in a secret and the rest in various
configmaps.

Change-Id: I3fe7e792e1febf2be2402d2e829c1262fd04f946