Store autohold requests in zookeeper
Storing autohold requests in ZooKeeper, rather than in-memory,
allows us to remember requests across restarts, and is a necessity
for future work to scale out the scheduler.

Future changes to build on this will allow us to store held node
information with the change for easy node identification, and to
delete any held nodes for a request using the zuul CLI.

A new 'zuul autohold-delete' command is added since hold requests
are no longer automatically deleted.

This makes the autohold API:
   zuul autohold: Create a new hold request
   zuul autohold-list: List current hold requests
   zuul autohold-delete: Delete a hold request

Change-Id: I6130175d1dc7d6c8ce8667f9b14ae9377737d280