DNM: Add script to pre-populate private keys
In some environments it is beneficial to have a master key in the
deployment tooling and derive all private keys from this
one. E.g. this eliminates the need of backing up secret data.

This adds a script that can do this and should be run before starting
and reconfiguring the scheduler.

Update:

This updates the reproducible private key generations script to use
ZooKeeper.  It also uses several Zuul library methods.

This is not intended as a long-term solution.  This is best used as
a transition to a system that exports keys from ZK for backup purposes.

Co-Authored-By: Tobias Henkel <tobias.henkel@bmw.de>
Change-Id: I562db318c18df67b4817875d6ea8517deeb7edf3