Add script to pre-populate private keys
In some environments it is beneficial to have a master key in the
deployment tooling and derive all private keys from this
one. E.g. this eliminates the need of backing up secret data.

This adds a script that can do this and should be run before starting
and reconfiguring the scheduler.

Change-Id: Ib5b3f423736864e69c72fd321b24ae702903f98e