Update patch set 1
Patch Set 1: Code-Review+2

This is cool. I think this should maybe be built-in to the scheduler, something like master_key=.... Using that method, we can store the master key in a KMS like Hashicorp Vault and remove it once it is read from disk. Then in this mode, the keys don't even go on disk, which is a win from a security perspective.

Patch-set: 1
Reviewer: Gerrit User 6488 <6488@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2