Update patch set 1
Patch Set 1:

Jeremy raised some good points in IRC about how pycrypto is abandonware, so I don't think we can add this to the scheduler as-is.  Maybe this "contrib" script is okay, though that's not ideal too.

I would love the functionality if we can get it in a supported manner.

Patch-set: 1