Update patch set 14
Patch Set 14: Code-Review-1

(1 comment)

As we now have allowed (limited) add_host to untrusted jobs I'm not sure about the big picture how this change fits into the architecture so I have a couple of open points.

This change restricts the feature to trusted jobs while limited dynamic add_host is already possible for untrusted jobs. I think we should have feature parity in this regard. This takes me to the question how we could achieve that without much code duplication. One way could be to just persist all added hosts in a special file within the add_host action plugin. But we'd need to override this for trusted jobs as well.

Another thing we should think about is whether we want this to be an api. Currently zuul_return only forwards generic variables to later jobs and not later playbooks in the same job. I think we should change this in any case. Then using this feature the user could just forward any information to the later playbooks and use this information to restore the added hosts via add_host.

Patch-set: 14