Update patch set 2
Patch Set 2:

That's no longer an experiment as it fix the random failure I see with the Pagure driver patch. But a double check from core is needed as I'm not sure why I got that random issue with that test (not related with the Pagure driver) and why the gc.disable() fix it.

Patch-set: 2
Label: Verified=0