Update patch set 8
Patch Set 8: Code-Review-1

This looks good, but setting -1 for now until the rest of the series in ready.  In particular, we still need:

* delete the held nodes when running autohold-delete
* docs for the new CLI commands (https://zuul-ci.org/docs/zuul/admin/client.html#autohold)
* release note

Today, when a user sets an autohold with an expiration (either explicitly or using the nodepool default hold expiration), it's a fire-and-forget situation.  It would be nice if we could keep that with this new system too.  Perhaps in some cases we can scan the autohold list and delete any requests whose held nodes no longer exist.  Or, perhaps when the last count is triggered for a request, we could set an expiration time on the request itself that is equal to the node expiration time.  Since that's when the nodes should be disappearing anyway, we shouldn't need the request longer than that.

Patch-set: 8
Label: Code-Review=-1