Update patch set 3
Patch Set 3:

It's possible this might be a little hard to use if someone makes a count>1 request for a multinode job because it may be difficult to see the multiple groupings.  However, that's no more difficult than what we have today, and it may not be worth the trouble to deal with.  We can always improve that later.

Patch-set: 3
Label: Verified=0