Update patch set 8
Patch Set 8:

> This looks good, but setting -1 for now until the rest of the
 > series in ready.  In particular, we still need:
 > 
 > * delete the held nodes when running autohold-delete

That change is in development now.

 > * docs for the new CLI commands (https://zuul-ci.org/docs/zuul/admin/client.html#autohold)
 > * release note

Yep, agreed these are needed.

 
 > Today, when a user sets an autohold with an expiration (either
 > explicitly or using the nodepool default hold expiration), it's a
 > fire-and-forget situation.  It would be nice if we could keep that
 > with this new system too.  Perhaps in some cases we can scan the
 > autohold list and delete any requests whose held nodes no longer
 > exist.  Or, perhaps when the last count is triggered for a request,
 > we could set an expiration time on the request itself that is equal
 > to the node expiration time.  Since that's when the nodes should be
 > disappearing anyway, we shouldn't need the request longer than
 > that.

Excellent point, and one that I hadn't considered until just now.

However, I question if we actually want to keep the fire-and-forget model going forward (we might, just playing devil's advocate). If we keep that, then we'll have two types of hold requests, dependent on node expiration: requests that must be manually deleted (no expiration), and requests that just "disappear".

Are the ones that just disappear a good admin experience? There would not be a way to easily figure out if a request was manually deleted by someone, or if it just expired. For example, an admin sets a hold request with node expiration of 1 week. While a dev is working on the problem, the node expiration hits and the node disappears before the dev finishes. The dev questions an admin on why it disappeared, but the admin no longer has the information to determine if the cause was a manual deletion or node expiration because the request is now gone from ZooKeeper.

Patch-set: 8
Label: Verified=0