Update patch set 8
Patch Set 8:

I think several of us opendev admins will confess to being lazy and not cleaning up held nodes, so the auto-expiration has been very good for that.  At this point, I don't even use any admin commands beyond setting the initial autohold (I use the job logs to find the ip of the held node).  So I'm fairly sure that if I had to manually remove the autohold request, I would forget, and we'd be back in the situation where Jens is asking if he can delete my old held node requests again :).

To address the audit concerns, here are two ideas: 1) log it and don't worry about it.  2) go with the idea of setting an expiration on the autohold request itself, but instead of making it the same as the node expiration, add 24h to it.  So the requests are still visible up to 24 hours after the nodes are cleaned up, but then disappear on their own.

Patch-set: 8