Safely add Ansible password lookup plugin
Ansible's password lookup plugin can be useful for generating random
passwords, but has other uses which could allow a nefarious user to
create new files or read the contents of existing files outside the
workspace (as long as those paths are writeable or readable by the
user under which Ansible is executing). To that end, short-circuit
and return an error if an untrusted use attempts to escape the
workspace in these ways.

Also increase the timeout for the plugins test by 50% since this
change increases the number of jobs in it by 42% and timeouts have
been seen as a result.

Change-Id: I0e372dda3a0f0f74d9e343c439514317dceb4d24