Revert "Revert "Create zuul/web/static on demand""
The js content tarball creation was broken. Instead of leaving the
symlink to a non-existing directory which gets created during python
setup we should just remove that symlink as well and create the
symlink and the static dir during the python setup. This way nothing
will be in the way of the javascript content generation.

This reverts commit eb7b18b38e01688b5b05e5dfdb359ff5d42d1ed8.

Change-Id: I5f8bfa62cd2d4d9823b86dbcda14885230847a82