Add spec for enhanced regional executor distribution
When running large distributed deployments it can be desirable to keep
traffic as local as possible. To facilitate this zuul supports zoning
of zuul-executors. Using zones executors only process jobs on nodes
that are running in the same zone. This works well in many
cases. However there is currently a limitation around live log
streaming that makes it impossible to use this feature in certain
environments.

Live log streaming via zuul-web or zuul-fingergw requires that each
executor is directly addressable from zuul-web or zuul-fingergw. This
is not the case if

* zuul-executors are behind a NAT. In this case one would need to
  create a NAT rule per executor on different ports which can become a
  maintenance nightmare.

* zuul-executors run in a different Kubernetes or OpenShift. In this
  case one would need a Ingress/Route or NodePort per executor which
  also makes maintenance really hard.

Change-Id: Iecff1b05a93b34c2a80d18df21eafdcc46f6fdfa