Update patch set 4
Patch Set 4: Code-Review-1

I understand the use case here, and I would like to figure out how to support it.  But I'm not comfortable with this change without some serious thought because it's very clearly a model violation.  We *intentionally* made it difficult to consult the event data in pipeline and job requirements because the triggering events are supposed to be completely independent from the contents of the queue.  That's the very thing that allows us to be able to say that a timer can enqueue a branch head in the first place.  Or that a message from IRC or the web app can enqueue a change.  The fact that this is only possible to implement because of some features we added for debugging purposes should be a red flag.

I'd like to see if it's possible to actually construct a coherent model of how this should work that doesn't involve the queue processor knowing anything about the event which triggered a change.

It's also worth thinking about whether this should be an attribute of the trigger.  Perhaps it should be a pipeline attribute (ie, pipeline.ignore-file-matchers).

Right now, the system does exactly what the user says.  If a user adds a file matcher to a job, we will not run the job unless the ref matches.  If people want to always run a job, then don't add a file matcher to it.  If they want to always run the job in a timer pipeline but not in a change pipeline, then only add the file matcher when adding it to the change pipeline (or make a parent job with no files matchers and a child job with).

Where this is useful is for very complex parent jobs with file matchers baked in which then have many child jobs in different repos where it would be inconvenient to duplicate the file matching.

If we added this as a pipeline attribute, it would allow us to say "it never makes sense for a job in this pipeline to consider file matching".  But we could still construct pipelines with timer triggers which *did* perform file matching.

Or, we could add a new job attribute which could be used to ignore file matchers, inverting the normal process, so that a job could be attached to a periodic pipeline while specifying that any file matchers from parent jobs should be ignored.

Honestly, I'm not sure what the right approach here is, but I think this warrants considerable thought, because this change fundamentally changes the contract between drivers, and whatever we do here could affect combinations we haven't thought about yet.

I see 3 options on the table:

A) This one -- specify at the trigger driver level whether file matchers should apply.
B) Specify at the pipeline level whether file matchers should apply.
C) Specify at the project-pipeline-job level whether file matchers should apply.

To evaluate (A) we should take all of the triggers we have today and can conceive of in the future and see whether we can definitively say whether changes from that trigger should always or never use file matchers.  Or would we expect individual events from some triggers to return different values for needsFileMatching()?

To evaluate (B) it's worth thinking about whether we can think of a pipeline which would have multiple triggers where events from one should ignore file matchers and events from another would not.

Option (C) should certainly work, but it's the least convenient for users.

To move forward with this, I think we should do the following:

1) The author and reviewers should consider the rambling (sorry!) statements above and see if we can think of any problems with (A), or whether we would prefer (B) or (C).

And if we stick with (A):

2) Please change the reno section from "fixes" to "features", as this is not a bug fix, but rather a behavior change.

3) Add documentation about the behavior to both the job.files attribute section in the docs as well as to the timer trigger.

Patch-set: 4
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1