Update patch set 4
Patch Set 4: -Code-Review

We should also consider option (D) which would be similar to option (A). We could make this decision from the type of the change instead of the event. E.g. a timer trigger always enqueues a branch change with the current branch tip which probably never needs file matching while a gerrit change or github pr probably always need file matching. Ref-Changes could make that decision a little bit smarter based on old-rev and new-rev.

If option (D) works it would have the advantage over (A) that it just works like (A) from the user's point of view while not having the problem with the model violation Jim mentioned.

Patch-set: 4