Update patch set 4
Patch Set 4:

The model is designed to give us the ability to have any trigger mechanism enqueue any change, and to have it always behave the same way.  If we violate that, we have to be very careful, otherwise we won't be able to do things like have the web trigger enqueue a gerrit change.

Of course we do have one example of data leakage -- ref-updated events create "Ref" objects with oldrev and newrev attributes.  That's blurring the line a bit, in that you can't necessarily reconstruct the oldrev attribute after the event has happened.  So anything that wants to enqueue a Ref with an oldrev has to supply that attribute.

I think that is drawing me to option (D) -- where we encode this information in the Ref object.  It's most like the current exception that we make for oldrev.  We'll have several kind of Ref objects:

Change (oldrev: no, files: yes)
Current ref: (oldrev: no, files: no)
Updated ref: (oldrev: yes, files: yes)

Then if there is a files list on the Ref, the files matcher will be active, if not, then it won't.

Patch-set: 4