[DNM][WIP] admin REST API: zuul-web integration
Add a configuration file allowing deployers to set up an OIDC provider
for authentication.
Save the authenticated user's JWT to the redux store for
consumption with the API.
If authenticated and allowed, add the following buttons to the GUI:

* autohold a failed build's resources on its next failure,
  from the builds page
* dequeue a buildset from the status page
* re-enqueue a buildset from the buildsets page

This is very much a WIP, use/review at your own risk!

Change-Id: I052c340b15b123ca387cff42aea5d09babe01404