admin REST API: zuul-web integration
Add a configuration file allowing deployers to set up an OIDC provider
for authentication. By default, no provider is configured and therefore
the GUI does not allow logging in/out nor access to privileged actions.

If a provider is configured, and a user is authenticated and allowed,
add the following buttons to the GUI:

* autohold a failed build's resources on its next failure,
  from the builds page
* dequeue a buildset from the status page
* re-enqueue a buildset from the buildsets page

This is a minimal, usable implementation. Until this feature is mature,
it is recommended to activate it only on "bleeding edge" deployments.

Change-Id: I052c340b15b123ca387cff42aea5d09babe01404