WIP: Remove fact cache restriction for localhost (zuul-executor)
When we originally created this logic, we didn't want to leak any
ansible facts on zuul-executors. However, as more jobs are written,
there are some facts that look to be helpful. In this case,
ansible_date_time will be collected and stored into localhost fact
cache.

* Need tests
* Does the break jobs that assumed * !all facts were primed into the
cache.
* Should we run 2, setup gathering tasks. Once for localhost (filtered),
one as normal.

Change-Id: Id2634fabb549257781fa7d40eee6c5bd34c37047
Signed-off-by: Paul Belanger <pabelanger@redhat.com>