Update patch set 53
Patch Set 53: Code-Review-1

One thing the series still needs before we can merge it is updates to the documentation.  We should be able to borrow from the spec to make that easier.  But we at least need to include the new tenant yaml config items, a general descriptive section about using and configuring auth/jwt, and also the new cli commands.

Patch-set: 53
Label: Code-Review=-1