Limit jobs runs with standard set of irrelevant files
Linting, unit tests, and other jobs were running even when only doc
files were changed. To limit what jobs get executed when there are no
functional code changes in a patch, this defines a standard set of files
to ignore that can be used for any jobs that would not need to be run.

Change-Id: Iefd91b727d59f85012ec09b9d2725e4432a20535
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>