Support ssl encrypted fingergw
When using fingergw for inter region log streaming it can be desirable
to support ssl encrypted connections with client auth just like we do
with gearman. This will also make it easy to route traffic to the
finger gateway via an openshift route using SNI and pass-through.

Docs and release note added in a subsequent change.

Change-Id: Ia5c739a3fcf229da140c4e2ebbe1a771c63b0489