WIP: Register finger gateway if it's zoned
TODO:
* docs
* tests

When we want to stream from an executor in a different zone via a
finger gateway in that zone we need a way to query for one in that
zone. Thus register the gearman function 'fingergw:info:<zone>' if the
fingergw is zoned. That way the scheduler will later be able to route
streams from different zones via finger gateways that are responsible
for their zone.

Change-Id: Ic5da5538af071297aa875d1e858ab41f1ffcdeac