Correct quick-start executor volume error
All zuul container images specify a volume at /var/lib/zuul,
therefore the docker-compose file does not need to specify the
same for the scheduler container.

This also means that we were inadvertently running the executor
with the git_dir and job_dir on different filesystems.  Since
we have configured (in the container images) /var/lib/zuul to
be a volume, we must ensure that job_dir is on that volume.
Update the zuul config to specify /var/lib/zuul/builds as the
job_dir.

Change-Id: I2fe5eea588006da7181c3ea8ad2637598764e8f1