Rework GitHub rate limit handling
Our current rate limit handling is simple and only logs the current
rate limit by calling the rate limit api. This will only ensure that
rate limit problems can be spotted in the logs. This reworks the rate
limit concept by installing a hook into the session that handles the
rate limit stuff completely transparent to the caller.

It performs rate limit logging based on the headers we get from github
so there are no further calls needed anymore to check the rate
limit.

Further it intercepts the 403 responses that are returned when we hit
the rate limit. Luckily GitHub tells how long we have to wait until we
can retry. The rate limit handler then waits for this time and retries
the request automatically. This way rate limit handling works in a
central place without the need to handle it everywhere we do api
requests.

Change-Id: I235a01283b609df95138a0983c5d38b1058f8dbd