Add central retry handling for github requests
Instead of spraying the code with retries on server errors we can
centralize the retry handling of GitHub api requests. This makes sense
for GET requests that returned 5xx server errors. This can be done by
attaching a response hook into the requests session that intercepts
those responses and retries them transparently. The retries are
handled with an exponential backoff.

Change-Id: Id4bef2720b8d164f06dc9f7bba3c8a468788eadf