Expose ansible_date_time instead of date_time
For user friendliness it's easier to expose the date_time as
ansible_date_time.

Without this patch, someone using the date fact would need
to template different variables, depending on whether the play
runs in promote or in check pipelines. This could lead to
things like "{{ (ansible_date_time | date_time)['time'] }}" which
is not really necessary.

This patch solves it by ensuring ansible_date_time is exposed
instead, and add extra tests to ensure the ansible_date_time
coming from Ansible is not used by default by checking the TZ.
This test can be bringing false positives if the server is
configured without a localtime or running in Reykjavík, but
I think it's acceptable for now.

Change-Id: Idedb5ae22d1f9ef8447796559b282e9f7cff7d91