Add repl server for debug purposes
Sometimes it can be necessary to debug complex problems in a running
instance. A repl server can be useful in this case. However this can
be highly dangerous in case the port is not correctly firewalled
of. Thus add an undocumented command to enable and disable it.

To enable:
zuul-scheduler repl

To disable
zuul-scheduler norepl

Change-Id: I860db5d088c6b07a66b4390cdfd66637341e7a7a
Co-Authored-By: Tobias Henkel <tobias.henkel@bmw.de>