Add repl server for debug purposes
Sometimes it can be necessary to debug complex problems in a running
instance. A repl server can be useful in this case. However this can
be highly dangerous in case the port is not correctly firewalled
of. Thus leave the switch off by default and undocumented as this
should only be used by experienced zuul developers who know what
they're doing.

Change-Id: I860db5d088c6b07a66b4390cdfd66637341e7a7a
Co-Authored-By: Tobias Henkel <tobias.henkel@bmw.de>