Update patch set 4
Patch Set 4: Code-Review+2

Ian, I'm unclear about the basis for your -1.  This patch, as I understand it, restores some functionality present in Ansible itself which we have previously excised from Zuul.  Here we are putting it back and it functions almost exactly the way it would in a normal Ansible situation (with the exception that we pretend the Ansible controller is always in UTC).

Your comments mostly applies to Ansible itself.  While I agree that the behavior of this variable in Ansible is not what anyone wants, Zuul is not the right place to correct that.

As for documentation, I don't think we need to re-document Ansible's behavior -- this is a standard Ansible fact.  In Zuul we have made many choices about what facts or modules to make available in the executor; we haven't documented any of them.

Patch-set: 4
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2