Defer setting build result to event queue
We currently set the result on the build object directly when we
receive the notification from the executor and afterwards enqueue a
BuildCompleted event to let the pipeline process this result. This was
good in the past because the processing in the pipeline didn't have an
impact on the scheduling of further jobs in the buildset. However
since we introduced the ability to skip child jobs via zuul_return
this assumption is no longer true. Now we have a race where we set the
build result, then the queue processor requests nodes for child jobs
and afterwards the completed event will be processed which skips the
child jobs. In this case we leak nodes of the child jobs that never
won't run.

This can be solved by deferring to set the result to the event queue
which eliminates the race.

Change-Id: I0631b97e6ae2bf7d883c6c6a189e9d41d99ece5d