Update patch set 4
Patch Set 4: -Code-Review

fair enough, I had in my mind we had documented what variables were exposed, but now I check again you're right they're not there.  so it's too much to expect this to start that and that was why I felt -1.

I would argue that ansible_date_time is actually fairly useless if you're writing anything vaguely generic/portable.  You never know what it's going to return and it's not monotonic.  It should be called "ansible_time_facts_ran" and might be useful to see how far out of date the cache is ... but not really for any sort of time stamp.

Using it means you see things like it works in CI when you've got a fresh run, but then fails in production when cached facts are used (e.g. the failure we had updating the timestamp in the DNS jobs).

So, my point was there's an argument to purposely omit this on "it's a gun pointed at your foot" arguments ... but I'm not sure there is appetite from anyone but me for that.

Patch-set: 4
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=0